hit good trap--bubble_sort

+++ nemu/src/monitor/sdb/sdb.c
@@ -142,9 +142,9 @@ static int cmd_help(char *args) {
 
 static int cmd_si(char *args){
     char *arg = strtok(NULL," ");
-    int n = 1;
+    uint64_t n = 1;
     if(arg){
-    sscanf(arg,"%d",&n);
+    sscanf(arg,"%ld",&n);
     }
     cpu_exec(n);
     return 0;
